% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_visualizer.R
\name{correlation_visualizer}
\alias{correlation_visualizer}
\title{Title}
\usage{
correlation_visualizer(id, cpm_values, clinical)
}
\arguments{
\item{id}{session ID}

\item{cpm_values}{cpm_table with symbols}

\item{clinical}{matrix with clinical data}
}
\value{

}
\description{
Title
}
