% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_visualizer.R
\name{correlation_visualizer_ui}
\alias{correlation_visualizer_ui}
\title{correlation_visualizer}
\usage{
correlation_visualizer_ui(id, clinical)
}
\arguments{
\item{id}{id for app}

\item{clinical}{clinical data}
}
\value{

}
\description{
correlation_visualizer
}
