% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reactome_page.R
\name{GO_page}
\alias{GO_page}
\title{Title}
\usage{
GO_page(id, data_reactome, participant, cpm_values, edgeR)
}
\arguments{
\item{id}{session id}

\item{data_reactome}{enrichment data (called reactome for convenience)}

\item{participant}{sample info}

\item{cpm_values}{cpm matrix with symbols column}

\item{edgeR}{edgeR matrix}
}
\value{

}
\description{
Title
}
