% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reactome_page.R
\name{GO_page_ui}
\alias{GO_page_ui}
\title{Title}
\usage{
GO_page_ui(id, data, edgeR)
}
\arguments{
\item{id}{session id}

\item{data}{GO data generated with CAMERA}

\item{edgeR}{DEG analysis generated with edgeR or limma}
}
\value{

}
\description{
Title
}
