% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reactome_page.R
\name{reactome_page}
\alias{reactome_page}
\title{Title}
\usage{
reactome_page(id, data_reactome, participant, cpm_values, edgeR)
}
\arguments{
\item{id}{session id}

\item{data_reactome}{reactome data generated by CAMERA}

\item{participant}{info on samples}

\item{cpm_values}{CPM_matrix}

\item{edgeR}{edgeR/limma data for DEG (called edgeR for convenience)}
}
\description{
Title
}
